Move Entity directive
