+++ src/VoxaApp.ts
@@ -28,6 +28,7 @@ import { Context as AWSLambdaContext } from "aws-lambda";
 import { LambdaLogOptions } from "lambda-log";
 import {
   Ask,
+  Entity,
   IDirective,
   IDirectiveClass,
   Reprompt,
@@ -136,7 +137,16 @@ export class VoxaApp {
     this.onAfterStateChanged(this.renderDirectives);
     this.onBeforeReplySent(this.saveSession, true);
 
-    this.directiveHandlers = [Say, SayP, Ask, Reprompt, Tell, Text, TextP];
+    this.directiveHandlers = [
+      Say,
+      SayP,
+      Ask,
+      Reprompt,
+      Tell,
+      Text,
+      TextP,
+      Entity,
+    ];
   }
 
   public validateConfig() {
