+++ src/platforms/alexa/AlexaPlatform.ts
@@ -28,7 +28,6 @@ import { OnSessionEndedError } from "../../errors";
 import { VoxaApp } from "../../VoxaApp";
 import { IVoxaEvent, IVoxaIntentEvent } from "../../VoxaEvent";
 import { IVoxaReply } from "../../VoxaReply";
-import { Entity } from "../shared";
 import { IVoxaPlatformConfig, VoxaPlatform } from "../VoxaPlatform";
 import { AlexaEvent } from "./AlexaEvent";
 import { AlexaReply } from "./AlexaReply";
@@ -134,7 +133,6 @@ export class AlexaPlatform extends VoxaPlatform {
       APLTTemplate,
       APLTCommand,
       StopAudio,
-      Entity,
     ];
   }
