+++ src/platforms/shared/directives.ts
@@ -1,62 +1,5 @@
-import { Response } from "ask-sdk-model";
 import * as _ from "lodash";
-import { AudioItem } from "virtual-alexa";
-import { IDirective } from "../../directives";
-import { ITransition } from "../../StateMachine";
 import { IVoxaEvent } from "../../VoxaEvent";
-import { IVoxaReply } from "../../VoxaReply";
-import { AlexaReply } from "../alexa/AlexaReply";
-import { DialogflowReply } from "../dialogflow/DialogflowReply";
-
-export class Entity implements IDirective {
-  public static key: string = "entities";
-  public static platform: string = "core";
-
-  public viewPath?: any | any[];
-
-  constructor(viewPath: any | any[]) {
-    this.viewPath = viewPath;
-  }
-
-  public async writeToReply(
-    reply: IVoxaReply,
-    event: IVoxaEvent,
-    transition?: ITransition,
-  ): Promise<void> {
-    let entity: any = this.viewPath;
-
-    const platform = _.get(event, "platform.name");
-
-    if (_.isString(this.viewPath)) {
-      entity = await event.renderer.renderPath(this.viewPath, event);
-    }
-
-    if (_.isPlainObject(entity)) {
-      entity = [entity];
-    }
-
-    if (!_.isArray(entity) || _.isEmpty(entity)) {
-      throw new Error(
-        "Please verify your entity it could be empty or is not an array",
-      );
-    }
-
-    entity = generateEntity(entity, event);
-
-    if (platform === "google") {
-      (reply as DialogflowReply).sessionEntityTypes = entity;
-    }
-
-    if (platform === "alexa") {
-      const response: Response = (reply as AlexaReply).response;
-      if (!response.directives) {
-        response.directives = [];
-      }
-
-      response.directives.push(...entity);
-    }
-  }
-}
 
 export enum EntityOverrideMode {
   Unspecified = "ENTITY_OVERRIDE_MODE_UNSPECIFIED",
